Moved fetchCardData function to Related.jsx to load all cards on initial load. Todo: code cleanup
